Code cleanup pre-release
